feat: Define "TrafficLight" class
TrafficLight class is a child class of TrafficObject.

The class has the public methods "void waitForGreen()" and
"void simulate()" as well as "TrafficLightPhase getCurrentPhase()",
where TrafficLightPhase is an enum that can be either "red" or "green".

Also, the class has a private method "void cycleThroughPhases()".

Furthermore, the class has a private member _currentPhase which takes
"red“ or "green" as its value.