feat: Implement a cycleThroughPhases method
Implement the function with an infinite loop that measures the time
between two loop cycles and toggles the current phase of the traffic
light between red and green.

The cycle duration is a random value between 4 and 6 seconds, and the
while-loop uses std::this_thread::sleep_for to wait 1ms between
two cycles.

The method is started in a thread from the thread pool in the base
class when the public method simulate is called.