feat: Implement the methods receive and waitForGreen
The method "receive" uses std::unique_lock<std::mutex> and
_condition.wait() to wait for and receive new messages and pull them
from the queue using move semantics. The received object is then
returned by the receive function.

The method "waitForGreen" is completed, in which an infinite
while loop runs and repeatedly calls the receive function on the
message queue. Once it receives TrafficLightPhase::green,
the method returns.